仕様変更を伴う大規模コミットです 一時不安定になる可能性があります
git-svn-id: http://springhead.info/spr2/Springhead2/trunk@5125 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDBox.h
@@ -10,7 +10,7 @@
 
 #include <SprCollision.h>
 #include <Foundation/Object.h>
-#include "CDConvex.h"
+#include <Collision/CDConvex.h>
 
 namespace Spr{;
 
@@ -64,7 +64,7 @@ public:
 	/// 直方体のサイズを取得
 	Vec3f GetBoxSize();
 	Vec3f* GetVertices();
-	CDFaceIf* GetFace(size_t i);
+	CDFaceIf* GetFace(int i);
 
 	/// 直方体のサイズを設定
 	Vec3f SetBoxSize(Vec3f boxSize);
